refactor(executors): port #6030 (codex half) — extract tool normalization from codex
Extract the hosted-tool passthrough set + free-plan tool gating
(normalizeCodexTools, isCodexFreePlan, CODEX_HOSTED_TOOL_TYPES) from
codex.ts into a pure leaf open-sse/executors/codex/tools.ts; codex.ts
re-exports them so external importers keep the "./codex.ts" path.

Ports only the codex half of upstream #6030. The base half
(base/reasoningEffort.ts) is a NO-PORT: the fork already extracted the
identical provider-aware reasoning-effort sanitation into
open-sse/executors/reasoningEffortMaps.ts (fork PR-014, Wave 6), so
importing upstream's parallel leaf would duplicate the module and produce
a duplicate `sanitizeReasoningEffortForProvider` binding in base.ts.

codex.ts 1427->1270. Split-guarded by tests/unit/codex-tools-split.test.ts.
